prelim version of solid-compatible simplycode

+++ base-components/raw-api/rawApi/delete.js
@@ -1,16 +1,10 @@
 function(endpoint, params={}) {
-  if (!params.token && editor.pageData.token) {
-    params.token = editor.pageData.token;
+  if (solidClientAuthentication.default.getDefaultSession().info.isLoggedIn) {
+    return solidClientAuthentication.default.fetch(simplyRawApi.url + endpoint + simplyRawApi.encodeGetParams(params), {
+      mode : 'cors',
+      headers: this.headers,
+      redirect: "manual",
+      method: "DELETE"
+    });
   }
-  if (params.token) {
-    this.headers['Authorization'] = "Bearer " + params.token;
-    delete params.token;
-  } else {
-    delete this.headers.Authentication;
-  }
-  return fetch(simplyRawApi.url + endpoint + "/" + simplyRawApi.encodeGetParams(params), {
-    mode : 'cors',
-    headers: this.headers,
-    method: "DELETE"
-  });
 }
